Add files via upload
The three .ino files are the Arduino codes for my base system and each sensor accordingly. On this version they work up until the point of going into the trigger stage of the system but without any UI being added. The libraries needed to run are not included in the repository