Add files via upload
            11/11/2020

For each sensor now the threshold selection works as it should. Also added current threshold value to the navigation menu.

            12/11/2020

Modified parts of the menu to make the menu more readable and easy. Largers texts where it was possible an convenient for the user.

            13/11/2020

Modified more parts of the menu to make the menu more readable and easy.